+++ CHANGELOG.md
@@ -4,3 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2019-08-14
+## Added
+- Evaluator classes migrations from OctoBot project
+- Matrix class that manage global evaluation dictionary
+- Matrix channel, producer and consumer
+- First API methods